+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.2.1] - 2021-02-10
+
+### Fixed
+Automatic tracking issue
+
 ## [1.2.0] - 2021-02-09
 
 ### Added